requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
imagehash>=4.3.0
requests>=2.31.0

# Evaluation & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking
wandb>=0.15.0

# Google Cloud
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.30.0

# Config
pyyaml>=6.0

# Notebook support
ipykernel>=6.25.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
